Update default registeredRepositories to use https transport
Since January 2020, Maven Central has switched to https and does not
allow http anymore. So update the built-in default URLs to Maven
Central to use https.
While we are at it, also use https for `maven.restlet.org`. The
generic domain is kept, instead of changing the URL to the now
official URL `https://maven.restlet.talend.com`.